requirements.txt
Django>=6.0
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
drf-yasg
django-filter
gunicorn
pytest
pytest-django
coverage
